Fix(runtime):windows build fail
Fix windows build fail, needs explicity convert and include precompiled header

--bug=81989185
Contributed-by:kalinshi

	modified:   hotspot/src/share/vm/gc_interface/collectedHeap.cpp
	modified:   hotspot/src/share/vm/services/gcHistogram.cpp

+++ hotspot/src/share/vm/gc_interface/collectedHeap.cpp
@@ -622,7 +622,7 @@ void CollectedHeap::update_minor_gc_frequency_histogram() {
        _last_minor_gc_time = os::javaTimeMillis();
      }else {
        size_t now = os::javaTimeMillis();
-       _minor_gc_frequency_histogram->add(now - _last_minor_gc_time);
+       _minor_gc_frequency_histogram->add((long)(now - _last_minor_gc_time));
        _last_minor_gc_time = now;
      }
    }
